+++ Changelog.md
@@ -10,6 +10,7 @@
   shown in bright red)
 * Emphasise most recent message in the log, show others grayed out
 * Mention items at player's position in the log
+* Improve the result of the charm monster spell
 * Remove the option to set the game difficulty from the command line
 * Increased robustness against broken configuration files